% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwmTimeseries_function.r
\name{dwmTimeseries}
\alias{dwmTimeseries}
\title{A Timeseries Function}
\usage{
dwmTimeseries(xData, y, m, d, aggFun)
}
\arguments{
\item{dataset, }{start year, months and day of the timeseries, function how the data should be combined (mean or sum)}
}
\description{
This function creates daily, weekly, monthly timeseries from daily input data
}
\examples{
dwmTimeseries(xData,y,m,d,aggFun)
}
\seealso{
\code{\link{zooreg}} which creates the timeseries
}

