% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensScatterPlot.R
\name{ensScatterPlot}
\alias{ensScatterPlot}
\title{A Timeseries Function}
\usage{
ensScatterPlot(predD, obsD, h, pointS, plotTitle, xlabel, ylabel, nRuns)
}
\description{
This function creates a scatter plot from time series data.
 @param predictions
 @param observations
 @param h
 @param pointS
 @param title of the plot, 
 @param labels for x- and 
 @param y-axis
 @param nRuns Ensemble size
 @export
 @examples 
 ensScatterPlot()
}

