% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCorrelation.R
\name{findCorrelation}
\alias{findCorrelation}
\title{Function for data acquisition. Find correlation in data set.}
\usage{
findCorrelation(input_data)
}
\arguments{
\item{time}{series}
}
\description{
Function for data acquisition. Find correlation in data set.
}
\examples{
findCorrelation(qQ.Mo)

}

