% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDischargeWithDischarge.R
\name{predictDischargeWithDischarge}
\alias{predictDischargeWithDischarge}
\title{Predict discharge with discharge}
\usage{
predictDischargeWithDischarge(data.input, start.date.input, nRuns, model.type,
  test.for.fit, error.test)
}
\arguments{
\item{data.input}{Timeseries generated with dwmTimeseries()}

\item{start.date.input}{Date, list with year, month, and day. e.g. c(2000,01,01)}

\item{nRuns}{Ensemble size}

\item{model.type}{Which model to run (currently only 'mlpe')}

\item{test.for.fit}{Which performance criteria to apply (currently only 'MAE')}

\item{error.test}{Some stuff ('MAE','RMSE)}
}
\description{
Function that predicts discharge with discharge.
}

