% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDischargeWithDischarge2.R
\name{predictDischargeWithDischarge2}
\alias{predictDischargeWithDischarge2}
\title{Predict discharge with discharge}
\usage{
predictDischargeWithDischarge2(data.input, start.date, qQ.mlpe, nRuns,
  model.type, test.for.fit, error.test)
}
\arguments{
\item{data.input}{Timeseries generated with dwmTimeseries()}

\item{start.date}{Date, list with year, month, and day. e.g. c(2000,01,01)}

\item{qQ.mlpe}{Fitted model saved previously}

\item{nRuns}{Ensemble size}

\item{model.type}{Which model to run (currently only 'mlpe')}

\item{test.for.fit}{Which performance criteria to apply (currently only 'MAE')}

\item{error.test}{Some stuff ('MAE','R22','RMSE')}
}
\description{
Function that predicts discharge with discharge.
}

