% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndProcessData.R
\name{readAndProcessData}
\alias{readAndProcessData}
\title{Read and process data.}
\usage{
readAndProcessData(data.input, start.date.input)
}
\arguments{
\item{data.input}{Time series to be processed.}

\item{start.date.input}{Date of the first entry of the time series. e.g. c(2000,01,01)}
}
\description{
Reads in time series date read from csv and aggregated values to monthly intervalls.
}

