% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensScatterPlot.R
\name{ensScatterPlot}
\alias{ensScatterPlot}
\title{A Timeseries Function}
\usage{
ensScatterPlot(predD, obsD, h, pointS, plotTitle, xlabel, ylabel, nRuns)
}
\arguments{
\item{h}{DESCRIPTION}

\item{pointS}{DESCRIPTION}

\item{nRuns}{Ensemble size}

\item{predictions}{DESCRIPTION}

\item{observations}{DESCRIPTION}

\item{title}{of the plot DESCRIPTION}

\item{labels}{for x- and DESCRIPTION}

\item{y-axis}{DESCRIPTION}
}
\description{
This function creates a scatter plot from time series data.
}
\examples{
ensScatterPlot() 
}

