% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensScatterPlot.R
\name{ensScatterPlot}
\alias{ensScatterPlot}
\title{A Timeseries Function}
\usage{
ensScatterPlot(predD, obsD, h, pointS, plotTitle, xlabel, ylabel)
}
\description{
This function creates a scatter plot from time series data.
 @param predictions, observations, h, pointS, title of the plot, labels for x- and y-axis
 @export
 @examples 
 ensScatterPlot()
}

